Huge refactor - separate those concerns baby
